Move faa_delays coordinator to its own file
